fixed inconsistent backprop
Former-commit-id: a960a1c50244d5ca6449caefebde17544b8e952c [formerly 8383c148871b1e73ee9f10da06e5c3cf68e7084a]
Former-commit-id: 0a73375896d6e1d6921caf00aaf1db9c2ba8ee9d

+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/transforms/Floor.java
@@ -25,6 +25,6 @@ public class Floor extends BaseElementWiseOp {
             return NDArrays.scalar(ComplexUtil.floor(c));
         }
         float val = (float) value;
-        return Math.floor(val);
+        return (float) Math.floor(val);
     }
 }
